+++ main.py
@@ -1,404 +1,4 @@
-import sys
-import os
-from datetime import datetime
-import uuid
-
-class Task:
-    def __init__(self, title, description=""):
-        self.id = str(uuid.uuid4())
-        self.title = title
-        self.description = description
-        self.created_at = datetime.now()
-        self.updated_at = datetime.now()
-        self.board = Board(f"Board: {title}", f"Board pour la tâche: {title}")
-
-    def get_board(self):
-        return self.board
-
-class List:
-    def __init__(self, title):
-        self.id = str(uuid.uuid4())
-        self.title = title
-        self.created_at = datetime.now()
-        self.tasks = []
-    
-    def add_task(self, task):
-        self.tasks.append(task)
-        return task
-    
-    def remove_task(self, task_id):
-        self.tasks = [t for t in self.tasks if t.id != task_id]
-        return True
-    
-class Board:
-    def __init__(self, title, description=""):
-        self.id = str(uuid.uuid4())
-        self.title = title
-        self.description = description
-        self.created_at = datetime.now()
-        self.updated_at = datetime.now()
-        self.lists = []
-        
-        self.add_list(List("À faire"))
-        self.add_list(List("En cours"))
-        self.add_list(List("Terminé"))
-    
-    def add_list(self, list_obj):
-        self.lists.append(list_obj)
-        self.updated_at = datetime.now()
-        return list_obj
-    
-    def remove_list(self, list_id):
-        self.lists = [l for l in self.lists if l.id != list_id]
-        self.updated_at = datetime.now()
-        return True
-    
-class Nest:
-    def __init__(self):
-        self.boards = []
-        self.current_board = None
-        self.navigation_stack = []
-
-    def create_board(self, title, description = ""):
-        board = Board(title, description)
-        self.boards.append(board)
-        return board
-    
-    def delete_board(self, board_id):
-        self.boards = [b for b in self.boards if b.id != board_id]
-        if self.current_board and self.current_board == board_id:
-            self.back_to_parent()
-        return True
-    
-    def select_board(self, board_id):
-        for board in self.boards:
-            if board.id == board_id:
-                self.current_board = board
-                return True
-        return False
-    
-    def get_current_board(self):
-        return self.current_board
-    
-    def navigate_to_task_board(self, list_id, task_id):
-        if not self.current_board:
-            return False
-        
-        for list_obj in self.current_board.lists:
-            if list_obj.id == list_id:
-                for task in list_obj.tasks:
-                    if task.id == task_id:
-                        self.navigation_stack.append((self.current_board, list_id, task_id))
-                        self.current_board = task.get_board()
-                        return True
-        return False
-    
-    def back_to_parent(self):
-        if not self.navigation_stack:
-            return False
-        
-        parent_info = self.navigation_stack.pop()
-        self.current_board = parent_info[0]
-        return True
-    
-    def add_list_to_current_board(self, title):
-        if not self.current_board:
-            return None
-        
-        list_obj = List(title)
-        self.current_board.add_list(list_obj)
-        return list_obj
-    
-    def remove_list_from_current_board(self, list_id):
-        if not self.current_board:
-            return False
-        
-        return self.current_board.remove_list(list_id)
-    
-    def add_task_to_list(self, list_id, title, description=""):
-        if not self.current_board:
-            return None
-        
-        for list_obj in self.current_board.lists:
-            if list_obj.id == list_id:
-                task = Task(title, description)
-                list_obj.add_task(task)
-                return task
-            
-        return None
-    
-    def move_task_between_lists(self, task_id, source_list_id, target_list_id):
-        if not self.current_board:
-            return False
-        
-        source_list = None
-        target_list = None
-
-        for list_obj in self.current_board.lists:
-            if list_obj.id == source_list_id:
-                source_list = list_obj
-            elif list_obj.id == target_list_id:
-                target_list = list_obj
-
-        if not source_list or not target_list:
-            return False
-        
-        task_to_move = None
-        for task in source_list.tasks:
-            if task.id == task_id:
-                task_to_move = task
-                break
-
-        if not task_to_move:
-            return False
-        
-        source_list.remove_task(task_id)
-        target_list.add_task(task_to_move)
-
-        return True
-        
-    def get_board_path(self):
-        path = []
-        for board_info in self.navigation_stack:
-            board = board_info[0]
-            path.append(board.title)
-
-        if self.current_board:
-            path.append(self.current_board.title)
-
-        return path
-
-class NestCLI:
-    def __init__(self):
-        self.app = Nest()
-        self.commands = {
-            "help": self.show_help,
-            "exit": self.exit_app,
-            "ls": self.list_items,
-            "create-board": self.create_board,
-            "select-board": self.select_board,
-            "add-list": self.add_list,
-            "add-task": self.add_task,
-            "open-task": self.open_task_board,
-            "back": self.go_back,
-            "move-task": self.move_task,
-            "path": self.show_path,
-            "clear": self.clear_screen
-        }
-
-    def clear_screen(self, *args):
-        os.system('cls' if os.name == 'nt' else 'clear')
-
-    def run(self):
-        print("=== Taskrioshka - CLI ===")
-        print("Tapez 'help' pour voir les commandes disponibles.")
-
-        while(True):
-            current = self.app.get_current_board()
-            prompt = "> " if not current else f"{current.title} > "
-
-            try:
-                command = input(prompt).strip()
-                if not command:
-                    continue
-
-                parts = command.split()
-                cmd = parts[0].lower()
-                args = parts[1:]
-
-                if cmd in self.commands:
-                    self.commands[cmd](*args)
-                else:
-                    print(f"Commande inconnue: {cmd}")
-                    print("Tapez 'help' pour voir les commandes disponibles.")
-
-            except Exception as e:
-                print(f"Erreur; {str(e)}")
-
-    def show_help(self, *args):
-        print("\n=== Commandes disponibles ===")
-        print("help                                 - Affiche cette aide")
-        print("exit                                 - Quitte l'application")
-        print("ls                                   - Liste les tableaux, listes et tâches")
-        print("create-board <titre> [description]   - Crée un nouveau tableau")
-        print("select-board <board_id>              - Sélectionne un tableau")
-        print("add-list <titre>                     - Ajoute une liste au tableau courant")
-        print("add-task <list_id> <titre> [desc]    - Ajoute une tâche à une liste")
-        print("open-task <list_id> <task_id>        - Ouvre le tableau d'une tâche")
-        print("back                                 - Retourne au tableau parent")
-        print("move-task <task_id> <src_id> <dst_id>- Déplace une tâche entre listes")
-        print("path                                 - Affiche le chemin de navigation")
-        print("clear                                - Efface l'écran\n")
-
-    def exit_app(self, *args):
-        print("au revoir!")
-        sys.exit(0)
-    
-    def list_items(self, *args):
-        if not args:
-            self._list_all()
-        elif args[0] == "boards":
-            self._list_boards()
-        elif args[0] == "lists" and len(args) > 1:
-            self._list_tasks(args[1])
-
-    def _list_all(self):
-        current = self.app.get_current_board()
-
-        if not current:
-            print("\n=== Tableaux principaux ===")
-            if not self.app.boards:
-                print("Aucun tableau créé. Utilisez 'create-board' pour commencer.")
-            else:
-                for idx, board in enumerate(self.app.boards):
-                    print(f"{idx+1}. [{board.id}] {board.title} - {board.description}")
-            print()
-            return
-        
-        print(f"\n=== Tableau: {current.title} ===")
-        print(f"Description: {current.description}")
-        
-        if not current.lists:
-            print("Aucune liste dans ce tableau. Utilisez 'add-list' pour en créer.")
-        
-        for idx, lst in enumerate(current.lists):
-            print(f"\n-- Liste {idx+1}: [{lst.id}] {lst.title} --")
-            if not lst.tasks:
-                print("  Aucune tâche dans cette liste.")
-            
-            for task_idx, task in enumerate(lst.tasks):
-                print(f"  {task_idx+1}. [{task.id}] {task.title}")
-                if task.description:
-                    print(f"     {task.description}")
-                print(f"     [Double-cliquez pour ouvrir le board imbriqué]")
-        print()
-
-    def _list_boards(self):
-        print("\n=== Tableaux principaux ===")
-        if not self.app.boards:
-            print("Aucun tableau créé.")
-        else:
-            for idx, board in enumerate(self.app.boards):
-                print(f"{idx+1}. [{board.id}] {board.title}")
-        print()
-
-    def _list_tasks(self, list_id):
-        current = self.app.get_current_board()
-        if not current:
-            print("Aucun tableau sélectionné.")
-            return
-        
-        for lst in current.lists:
-            if lst.id == list_id:
-                print(f"\n=== Tâches dans la liste: {lst.title} ===")
-                if not lst.tasks:
-                    print("Aucune tâche dans cette liste.")
-                
-                for idx, task in enumerate(lst.tasks):
-                    print(f"{idx+1}. [{task.id}] {task.title}")
-                    if task.description:
-                        print(f"   {task.description}")
-                print()
-                return
-        
-        print(f"Liste avec ID {list_id} non trouvée.")
-
-    def create_board(self, *args):
-        if not args:
-            print("Erreur: Le titre du tableau est requis.")
-            return
-        
-        title = args[0]
-        description = " ".join(args[1:]) if len(args) > 1 else ""
-        
-        board = self.app.create_board(title, description)
-        print(f"Tableau créé avec succès: [{board.id}] {board.title}")
-
-    def select_board(self, *args):
-        if not args:
-            print("Erreur: L'ID du tableau est requis.")
-            return
-        
-        board_id = args[0]
-        if self.app.select_board(board_id):
-            board = self.app.get_current_board()
-            print(f"Tableau sélectionné: {board.title}")
-        else:
-            print(f"Tableau avec ID {board_id} non trouvé.")
-
-    def add_list(self, *args):
-        if not args:
-            print("Erreur: Le titre de la liste est requis.")
-            return
-        
-        current = self.app.get_current_board()
-        if not current:
-            print("Erreur: Aucun tableau sélectionné.")
-            return
-        
-        title = " ".join(args)
-        list_obj = self.app.add_list_to_current_board(title)
-        print(f"Liste ajoutée: [{list_obj.id}] {list_obj.title}")
-
-    def add_task(self, *args):
-        if len(args) < 2:
-            print("Erreur: L'ID de la liste et le titre de la tâche sont requis.")
-            return
-        
-        list_id = args[0]
-        title = args[1]
-        description = " ".join(args[2:]) if len(args) > 2 else ""
-        
-        task = self.app.add_task_to_list(list_id, title, description)
-        if task:
-            print(f"Tâche ajoutée: [{task.id}] {task.title}")
-        else:
-            print(f"Liste avec ID {list_id} non trouvée.")
-    
-    def open_task_board(self, *args):
-        if len(args) < 2:
-            print("Erreur: L'ID de la liste et l'ID de la tâche sont requis.")
-            return
-        
-        list_id = args[0]
-        task_id = args[1]
-        
-        if self.app.navigate_to_task_board(list_id, task_id):
-            board = self.app.get_current_board()
-            print(f"Navigué vers le tableau de la tâche: {board.title}")
-        else:
-            print("Tâche ou liste non trouvée.")
-    
-    def go_back(self, *args):
-        if self.app.back_to_parent():
-            board = self.app.get_current_board()
-            print(f"Retourné au tableau parent: {board.title}")
-        else:
-            print("Déjà au niveau principal.")
-    
-    def move_task(self, *args):
-        if len(args) < 3:
-            print("Erreur: L'ID de la tâche, l'ID de la liste source et l'ID de la liste cible sont requis.")
-            return
-        
-        task_id = args[0]
-        source_list_id = args[1]
-        target_list_id = args[2]
-        
-        if self.app.move_task_between_lists(task_id, source_list_id, target_list_id):
-            print(f"Tâche [{task_id}] déplacée avec succès.")
-        else:
-            print("Échec du déplacement. Vérifiez les IDs.")
-    
-    def show_path(self, *args):
-        path = self.app.get_board_path()
-        if not path:
-            print("Vous êtes au niveau principal.")
-            return
-        
-        print("Chemin de navigation:")
-        print(" > ".join(path))
-
+from taskrioshka.views.cli.nest_cli import NestCLI
 
 if __name__ == "__main__":
     cli = NestCLI()
