+++ taskrioshka/__init__.py
@@ -1,0 +1,4 @@
+""" Taskrioshka. """
+
+__version__ = '0.1.0'
+__author__ = 'TournoisAlexandre'
