+++ taskrioshka/views/cli/__init__.py
@@ -1,0 +1,1 @@
+"""CLI"""
